requirements.txt
opencv-python
numpy
torch
ultralytics
simple-lama-inpainting
pillow